collapse into evaluation.py
